untuk fungsi cart sudah sesuai jika ada id yang sama maka tidak terinputkan serta custom text sesuai dengan kondisi, memperbaiki posisi linkserver dipindahkan ke state index.js untuk mempermudah, menambah return pada book.js

+++ src/store/modules/transaction.js
@@ -4,6 +4,7 @@ export const namespaced = true
 
 export const state = {
     cart: [],
+    text: null,
     linkServer: 'http://127.0.0.1:8000/'
 }
 
@@ -27,8 +28,29 @@ export const actions = {
     //         })
     // },
     sendToCart({commit,state},item){
-        commit('SET_TO_CART',item)
-        localStorage.setItem('book-cart',JSON.stringify(state.cart))
+        //if state.cart still empty
+        if(state.cart.length == 0){
+            commit('SET_TO_CART',item)
+            localStorage.setItem('book-cart',JSON.stringify(state.cart))
+        }
+        //if not
+        else {
+            //check if id exist
+            //using some will return true or false when find matching item
+            const existId = state.cart.some(el => el.id === item.id)
+
+            //if true 
+            if (existId){
+                state.text = 'Sudah pernah ditambahkan ke keranjang'
+            }
+            //if false
+            else {
+                commit('SET_TO_CART',item)
+                state.text = 'Berhasil ditambahkan ke keranjang'
+                localStorage.setItem('book-cart',JSON.stringify(state.cart))
+            }
+        }
+        
     }
 }
